+++ app/filters/is_admin.py
@@ -11,4 +11,4 @@ class AdminFilter(BoundFilter):
         self.is_admin = is_admin
 
     async def check(self, message: types.Message):
-        return message.from_user.id in Config.admins
+        return message.from_user.id in Config.ADMINS
