+++ app/handlers/admins/__init__.py
@@ -1,4 +1,4 @@
-from aiogram import Dispatcher
+from aiogram import Dispatcher, types
 
 from app.config import Config
 from app.handlers.admins.admin import broadcast, cancel_broadcast, start_broadcasting
@@ -6,6 +6,7 @@ from app.states.admin_states import BroadcastAdmin
 
 
 def setup_admin(dp: Dispatcher):
-    dp.register_message_handler(broadcast, commands=["broadcast"], user_id=Config.admins)
-    dp.register_callback_query_handler(cancel_broadcast, text='cancel', state='*', user_id=Config.admins)
-    dp.register_message_handler(start_broadcasting, state=BroadcastAdmin.BROADCAST, user_id=Config.admins)
+    dp.register_message_handler(broadcast, commands=["broadcast"], user_id=Config.ADMINS)
+    dp.register_callback_query_handler(cancel_broadcast, text='cancel', state='*', user_id=Config.ADMINS)
+    dp.register_message_handler(start_broadcasting, state=BroadcastAdmin.BROADCAST, user_id=Config.ADMINS,
+                                content_types=types.ContentType.ANY)
