+++ app/middlewares/__init__.py
@@ -2,7 +2,7 @@ from aiogram import Dispatcher
 from aiogram.contrib.middlewares.environment import EnvironmentMiddleware
 from aiogram.contrib.middlewares.logging import LoggingMiddleware
 
-from .data import DataMiddleware
+from .acl import ACLMiddleware
 from .database import DatabaseMiddleware
 from .throttling import ThrottlingMiddleware
 
@@ -11,5 +11,5 @@ def setup(dp: Dispatcher):
     dp.setup_middleware(LoggingMiddleware())
     dp.setup_middleware(EnvironmentMiddleware())
     dp.setup_middleware(DatabaseMiddleware())
-    dp.setup_middleware(DataMiddleware())
+    dp.setup_middleware(ACLMiddleware())
     dp.setup_middleware(ThrottlingMiddleware())
