+++ app/middlewares/throttling.py
@@ -42,8 +42,8 @@ class ThrottlingMiddleware(BaseMiddleware):
             key = f"{self.prefix}_message"
         delta = throttled.rate - throttled.delta
         if throttled.exceeded_count <= 2:
-            await message.reply('Слишком много запросов!')
+            await message.reply('Too many requests!')
         await asyncio.sleep(delta)
         thr = await dispatcher.check_key(key)
         if thr.exceeded_count == throttled.exceeded_count:
-            await message.reply('Разблокировано.')
+            await message.reply('Unblocked.')
